\docType{package}
\name{afe-package}
\alias{afe}
\alias{afe-package}
\title{Analysis of Factorial Experiments.}
\description{
  Analysis of Factorial Experiments.
}
\details{
  \tabular{ll}{ Package: \tab afe\cr Type: \tab Package\cr
  Version: \tab 0.1-17\cr Date: \tab 2012-08-07\cr Depends:
  \tab R (>= 2.14.0), car, reshape2, stringr\cr Encoding:
  \tab UTF-8\cr License: \tab GPL (>=3)\cr }

  Provides convenience functions for analyzing factorial
  experiments, namely provide convenience tools for
  repeated-measures ANOVA and ANCOVA analysis with type 2
  and type 3 sums of squares wrapping car::Anova().
}
\author{
  Henrik Singmann
  \email{henrik.singmann@psychologie.uni-freiburg.de}
}
\keyword{package}

